Move VirtualType enum from annotations to virtual_registers
  This resolves a layering issue where virtual_registers depended on
  annotations just for the VirtualType enum. The enum now lives in
  virtual_registers.virtual_type, and annotations.virtual_type re-exports
  it alongside TWIRLING_GROUPS and GroupLiteral for backward compatibility.
