+++ samplomatic/annotations/__init__.py
@@ -17,4 +17,4 @@ from .dressing_mode import DressingLiteral, DressingMode
 from .inject_local_clifford import InjectLocalClifford
 from .inject_noise import InjectionSite, InjectNoise
 from .twirl import Twirl
-from .virtual_type import TWIRLING_GROUPS, GroupLiteral, VirtualType
+from .virtual_type import TWIRLING_GROUPS, GroupLiteral
