+++ samplomatic/samplex/nodes/multiplication_node.py
@@ -13,9 +13,8 @@
 """MultiplicationNode"""
 
 from ...aliases import RegisterName
-from ...annotations import VirtualType
 from ...exceptions import SamplexConstructionError
-from ...virtual_registers import GroupRegister, VirtualRegister
+from ...virtual_registers import GroupRegister, VirtualRegister, VirtualType
 from .evaluation_node import EvaluationNode
