+++ samplomatic/samplex/nodes/slice_register_node.py
@@ -18,9 +18,8 @@ import numpy as np
 from numpy.typing import ArrayLike
 
 from ...aliases import RegisterName, SubsystemIndex
-from ...annotations import VirtualType
 from ...exceptions import SamplexConstructionError
-from ...virtual_registers import VirtualRegister
+from ...virtual_registers import VirtualRegister, VirtualType
 from .evaluation_node import EvaluationNode
