+++ samplomatic/virtual_registers/c1_register.py
@@ -18,10 +18,10 @@ import numpy as np
 from numpy.typing import ArrayLike
 from qiskit.quantum_info import Clifford
 
-from ..annotations import VirtualType
 from .finite_group_register import FiniteGroupRegister
 from .tables.c1_tables import C1_INVERSE_TABLE, C1_LOOKUP_TABLE
 from .u2_register import U2Register
+from .virtual_type import VirtualType
 
 C1_TO_TABLEAU = np.array(
     [
