+++ samplomatic/virtual_registers/virtual_type.py
@@ -1,6 +1,6 @@
 # This code is a Qiskit project.
 #
-# (C) Copyright IBM 2025.
+# (C) Copyright IBM 2025-2026.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,11 +10,22 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+"""VirtualType"""
 
-from .c1_register import C1Register
-from .group_register import GroupRegister
-from .pauli_register import PauliRegister
-from .u2_register import U2Register
-from .virtual_register import VirtualRegister
-from .virtual_type import VirtualType
-from .z2_register import Z2Register
+from enum import Enum
+
+
+class VirtualType(str, Enum):
+    """Name identifiers for virtual gate types."""
+
+    PAULI = "pauli"
+    """The projective Pauli group."""
+
+    C1 = "c1"
+    """The single-qubit Clifford group."""
+
+    U2 = "u2"
+    r"""The :math:`2 \times 2` unitary group."""
+
+    Z2 = "z2"
+    """The two-element cyclic group."""
