+++ test/unit/test_distributions/test_haar_u2.py
@@ -14,8 +14,8 @@
 
 import numpy as np
 
-from samplomatic.annotations import VirtualType
 from samplomatic.distributions import HaarU2, UniformPauli
+from samplomatic.virtual_registers import VirtualType
 
 
 def test_attributes():
